cleanup: types (#254)
* cleanup: types

* move
