+++ packages/core/base/src/index.ts
@@ -1,10 +1,3 @@
-export * from "./utils/index";
-export * from "./models/types";
-export * from "./models/paymentElement";
-export * from "./models/events";
-export * from "./models/errors";
-export * from "./services/crossmintModalService";
-export * from "./services/crossmintPayButtonService";
-export * from "./services/crossmintPaymentService";
-export * from "./services/useCrossmintEvents";
-export * from "./services/crossmintUiService";
+export * from "./utils";
+export * from "./types";
+export * from "./services";
