--- packages/core/base/src/models/paymentElement.ts
@@ -1,155 +1,0 @@
-import { CheckoutEvents, CrossmintEventError, CrossmintEventErrorPayload, PaymentElementSDKEvents } from "./events";
-import { CaseInsensitive, CollectionId, Currency, Locale, PaymentMethod, UIConfig } from "./types";
-
-export type Recipient = {
-    email?: string;
-    wallet?: string;
-};
-
-export interface CrossmintCheckoutEvent<K extends CheckoutEvents = CheckoutEvents> {
-    type: K;
-    payload: CheckoutEventMap[K];
-}
-export type CrossmintCheckoutEventUnion = {
-    [K in CheckoutEvents]: CrossmintCheckoutEvent<K>;
-}[CheckoutEvents];
-
-export type MintConfig = Record<string, any> | Record<string, any>[];
-
-export type PaymentElementExperimentalOptions = {
-    useCardWalletEmail?: boolean; // Allows the recipient's email to be obtained from the card wallet. Currently only supported when `recipient.wallet` is also passed. If an email is already specified in `recipient.email` or typed by the user, it will take precedence over the email obtained from the card wallet.
-};
-
-// TODO: Remmeber to update this same interface in the Vue component aswell.
-// packages/ui/vue-ui/src/components/CrossmintPaymentElement.vue
-export type PaymentElement = {
-    projectId?: string;
-    mintConfig?: MintConfig;
-    recipient?: Recipient;
-    paymentMethod?: PaymentMethod;
-    currency?: CaseInsensitive<Currency>;
-    locale?: Locale;
-    uiConfig?: UIConfig;
-    environment?: string;
-    whPassThroughArgs?: any;
-    onEvent?(event: CrossmintCheckoutEventUnion): void;
-    cardWalletPaymentMethods?: CardWalletPaymentMethod | CardWalletPaymentMethod[] | "none";
-    emailInputOptions?: EmailInputOptions;
-    experimental?: PaymentElementExperimentalOptions;
-} & CollectionId;
-
-export type EmailInputOptions =
-    | {
-          show: true;
-          useStripeLink?: boolean;
-      }
-    | {
-          show: false;
-      };
-
-export type CardWalletPaymentMethod = "apple-pay" | "google-pay";
-
-export interface FiatPrice {
-    amount: string;
-    currency: Currency;
-}
-
-export interface LineItemMetadata {
-    imageUrl?: string;
-    description?: string;
-    title?: string;
-    collection?: string;
-}
-
-export interface LineItem {
-    price: FiatPrice;
-    gasFee?: FiatPrice;
-    metadata: LineItemMetadata;
-    quantity: number;
-}
-
-export interface InitialQuoteOutput {
-    totalPrice: FiatPrice;
-    lineItems: LineItem[];
-}
-
-interface PaymentRejectedPayload extends CrossmintEventErrorPayload {
-    orderIdentifier: string;
-    paymentMethodType: PaymentMethodType;
-}
-
-type EmptyObject = Record<string, never>;
-
-interface PaymentCompletedPayload {
-    orderIdentifier: string;
-}
-
-interface TransactionBase {
-    transactionIdentifier: string;
-    price: FiatPrice;
-}
-
-interface EvmTransaction {
-    contractAddress: string;
-    tokenIds: string[];
-}
-
-interface SolanaTransaction {
-    mintHash: string;
-}
-
-type TransactionFulfillmentSucceededPayload = TransactionBase & {
-    txId: string;
-} & (EvmTransaction | SolanaTransaction);
-
-interface TransactionFulfillmentFailed extends TransactionBase {
-    error: CrossmintEventError;
-}
-
-type Verification = { required: false } | { required: true; url: string };
-
-export const PaymentMethodType = {
-    // TODO: Think granularity for crypto payments.
-    CREDIT_CARD: "credit-card",
-    GOOGLE_PAY: "google-pay",
-    APPLE_PAY: "apple-pay",
-} as const;
-export type PaymentMethodType = (typeof PaymentMethodType)[keyof typeof PaymentMethodType];
-
-interface OrderProcessFinished {
-    successfulTransactionIdentifiers: string[];
-    failedTransactionIdentifiers: string[];
-    verification: Verification;
-    totalPrice: FiatPrice;
-    paymentMethodType: PaymentMethodType;
-}
-
-interface RecipientWalletChanged {
-    wallet: string;
-}
-interface RecipientEmailChanged {
-    email: string;
-}
-
-export interface CheckoutEventMap {
-    [CheckoutEvents.PAYMENT_PREPARATION_SUCCEEDED]: EmptyObject;
-    [CheckoutEvents.QUOTE_STATUS_CHANGED]: InitialQuoteOutput;
-    [CheckoutEvents.QUOTE_STATUS_INVALIDATED]: EmptyObject;
-    [CheckoutEvents.PAYMENT_PROCESS_STARTED]: EmptyObject;
-    [CheckoutEvents.PAYMENT_PREPARATION_FAILED]: CrossmintEventErrorPayload;
-    [CheckoutEvents.PAYMENT_PROCESS_SUCCEEDED]: PaymentCompletedPayload;
-    [CheckoutEvents.PAYMENT_PROCESS_CANCELED]: EmptyObject;
-    [CheckoutEvents.PAYMENT_PROCESS_REJECTED]: PaymentRejectedPayload;
-    [CheckoutEvents.ORDER_PROCESS_STARTED]: EmptyObject;
-    [CheckoutEvents.TRANSACTION_FULFILLMENT_SUCCEEDED]: TransactionFulfillmentSucceededPayload;
-    [CheckoutEvents.TRANSACTION_FULFILLMENT_FAILED]: TransactionFulfillmentFailed;
-    [CheckoutEvents.ORDER_PROCESS_FINISHED]: OrderProcessFinished;
-    [CheckoutEvents.RECIPIENT_WALLET_CHANGED]: RecipientWalletChanged;
-    [CheckoutEvents.RECIPIENT_EMAIL_CHANGED]: RecipientEmailChanged;
-}
-
-export type ParamsUpdatePayload = Partial<Record<keyof Omit<PaymentElement, "onEvent" | "environment">, any>>;
-
-export interface SDKEventMap {
-    [PaymentElementSDKEvents.PARAMS_UPDATE]: ParamsUpdatePayload;
-}
