+++ packages/core/base/src/types/errors.ts
