+++ packages/core/base/src/utils/i18n/index.ts
@@ -1,20 +1,19 @@
-import { Locale } from "../../models/types";
-import { NestedPaths, TypeFromPath } from "../../models/system";
-import zhCN from "./dictionaries/zhCN";
-import itIT from "./dictionaries/itIT";
-import enUS from "./dictionaries/enUS";
-import frFR from "./dictionaries/frFR";
-import zhTW from "./dictionaries/zhTW";
-import trTR from "./dictionaries/trTR";
-import deDE from "./dictionaries/deDE";
-import esES from "./dictionaries/esES";
-import ruRU from "./dictionaries/ruRU";
-import ukUA from "./dictionaries/ukUA";
-import koKR from "./dictionaries/koKR";
+import { Locale } from "../../types";
+import { NestedPaths, TypeFromPath } from "../../types/system";
 import Klingon from "./dictionaries/Klingon";
+import deDE from "./dictionaries/deDE";
+import enUS from "./dictionaries/enUS";
+import esES from "./dictionaries/esES";
+import frFR from "./dictionaries/frFR";
+import itIT from "./dictionaries/itIT";
+import koKR from "./dictionaries/koKR";
 import ptPT from "./dictionaries/ptPT";
+import ruRU from "./dictionaries/ruRU";
 import thTH from "./dictionaries/thTH";
-
+import trTR from "./dictionaries/trTR";
+import ukUA from "./dictionaries/ukUA";
+import zhCN from "./dictionaries/zhCN";
+import zhTW from "./dictionaries/zhTW";
 
 const localeMap = {
     "en-US": enUS,
@@ -30,7 +29,7 @@ const localeMap = {
     "tr-TR": trTR,
     "uk-UA": ukUA,
     "th-TH": thTH,
-    "Klingon": Klingon,
+    Klingon: Klingon,
 };
 
 export function t<K extends NestedPaths<typeof enUS>>(wordingKey: K, locale: Locale): TypeFromPath<typeof enUS, K> {
